requirements.txt
fastapi==0.110.0
pillow
requests
uvicorn[standard]